// src/users/users.controller.interface.ts
import { Request, Response, NextFunction } from "express";

export interface IUserController {
    login: (req: Request, res: Response, next: NextFunction) => void;
    register: (req: Request, res: Response, next: NextFunction) => void;
    info: (req: Request, res: Response, next: NextFunction) => void;
}